+++ packages/sdk-packs/sdk-57/template/app/_layout.tsx
@@ -4,18 +4,21 @@ import { StatusBar } from "expo-status-bar";
 import { DataProvider } from "../src/data/provider";
 import { OnboardingProvider } from "../src/onboarding/provider";
 import { SessionProvider } from "../src/session/provider";
+import { ThemeProvider } from "../src/theme/provider";
 
 export default function RootLayout() {
   return (
     <>
       <StatusBar style="auto" />
-      <SessionProvider>
-        <DataProvider>
-          <OnboardingProvider>
-            <Stack screenOptions={{ headerShown: false }} />
-          </OnboardingProvider>
-        </DataProvider>
-      </SessionProvider>
+      <ThemeProvider>
+        <SessionProvider>
+          <DataProvider>
+            <OnboardingProvider>
+              <Stack screenOptions={{ headerShown: false }} />
+            </OnboardingProvider>
+          </DataProvider>
+        </SessionProvider>
+      </ThemeProvider>
     </>
   );
 }