+++ packages/expojet/CHANGELOG.md
@@ -1,5 +1,12 @@
 # expojet
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [create-expojet@0.6.1]
+  - create-expojet@0.6.1
+
 ## 0.6.0
 
 ### Minor Changes